pyproject.toml
[project]
name = "enhance"
version = "0.2.0"
description = "Enhance AI - Sharpen, Denoise, and Upscale images using community models."
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "deferred-import>=0.1.0",
    "huggingface-hub>=0.30.2",
    "jsonpickle>=4.0.5",
    "numpy",
    "nvidia-ml-py>=13.580.82",
    "opencv-python-headless>=4.11.0.86",
    "pillow>=11.2.1",
    "psutil>=7.1.3",
    "pyside6>=6.9.0",
    "spandrel>=0.4.1",
    "spandrel-extra-arches>=0.2.0",
    "tifftools>=1.6.0",
    "sam2>=1.1.0",
    "transformers>=4.57.3",
    "timm>=1.0.22",
]

[project.optional-dependencies]
cpu = [
  "torch>=2.6.0",
  "torchvision",
]
cu118 = [
  "torch>=2.6.0",
  "torchvision",
]
cu126 = [
  "torch>=2.6.0",
  "torchvision",
]
mps = [
  "torch>=2.6.0",
  "torchvision",
]

[tool.uv]
no-build-isolation = true
conflicts = [
  [
    { extra = "cpu" },
    { extra = "cu118" },
    { extra = "cu126" },
    { extra = "mps" },
  ],
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu118", extra = "cu118" },
  # TODO: UV fails to resolve path for torch+cu126 on Windows. Use cu124 index.
  { index = "pytorch-cu124", extra = "cu126", marker = "platform_system == 'Windows'" },
  { index = "pytorch-cu126", extra = "cu126", marker = "platform_system != 'Windows'" },
  { index = "pytorch-cpu", extra = "mps" },
]
sam-2 = { git = "https://github.com/facebookresearch/sam2.git", rev = "c2ec8e14a185632b0a5d8b161928ceb50197eddc" }

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

